Fix mistake in marray && and || operator description and formatting
This commit removes the restriction on types for the && and || operator
when the left operand is a convertible scalar data type. This is in line
with the other variants of this operator. Likewise it fixes the return
type of the operator.

Additionally, this commit fixes the indentation of part of the marray
declaration header.

Signed-off-by: Larsen, Steffen <steffen.larsen@intel.com>